A few CSS changes (removing images) and fixed alpha/taxonomic view change and display issues:
The listing is handled by a CollectionView that wraps each item in a div
and applie the appropriate class. For alphabetical view, the taxa bar
items simple have no content but are still iterated over by the
CollectionView. The solution in this case is to hide the elements --
this shouldn't be a performance issue as the number of DOM elements is
insignificant and they will never be rendered anyway. Better solutions
welcome!