Merged search routes in to a single route with parameters
A possible bug in IE means that the search_without_render route took
precendence over the search route, so pages loaded without rendering the
filter controls. The routes have now been combined and take a "redraw"
option that decides whether or not to redraw the filter controls
(defaults to true).